cleanup Integrals_Monoelec for complex AOs

+++ src/Integrals_Monoelec/ao_mono_ints.irp.f
@@ -1,5 +1,5 @@
- BEGIN_PROVIDER [ double precision, ao_mono_elec_integral,(ao_num,ao_num)]
-&BEGIN_PROVIDER [ double precision, ao_mono_elec_integral_diag,(ao_num)]
+ BEGIN_PROVIDER [ complex*16, ao_mono_elec_integral,(ao_num,ao_num)]
+&BEGIN_PROVIDER [ complex*16, ao_mono_elec_integral_diag,(ao_num)]
   implicit none
   integer :: i,j,n,l
   BEGIN_DOC
